==> Code/3_preprocess_video.ipynb <==
import glob
import os
import subprocess
import numpy as np
import pandas as pd
import math
import time

from tqdm import tqdm
from pathlib import Path

import re

# What's happening here?

In this script we will load in all the video's and analyze them using OpenFace. 
The starting point of the experiment will be read in manually and this will be used to sync the video with the behavioral data. 

When the behavioral data is loaded in, a subindex for each trial can be made. 
On this information, averages will be run

This information will be stored in a df
- Resulting in a ppt x trial dataframe
-- Also containing every mean action unit for the duration of exposure ++ Date (from file name)



# Set Paths

# Declare all directories:

data_dir = Path('Z:/shares/ghepmk_data/2020_Kappen_PMS')

# Create all relevant directories
date_dir = "24082021/"
data_all_dir = os.path.join(data_dir, date_dir)
data_A_dir = os.path.join(data_dir, Path('10001'))
data_B_dir = os.path.join(data_dir, Path('10002'))

# Testmoment A

os.chdir(data_A_dir)
AVid = glob.glob("video_{ParticipantNo*" + "*.webm", recursive = True) # get all webm's in directory
AVidID = []
CSVy_n = []

for index, item in enumerate(AVid):
    AVidID.append(AVid[index].split("video_{ParticipantNo}_")[1][0:3])
        
AVidID = set(AVidID)
demographics = pd.read_csv(data_all_dir + "cleanData.csv", encoding = "ISO-8859-1")

exclude = ['&à&', # Weird shit
           '667', # Two files, 16-7 and 12-8. See also behavioral file
           '546', # Two files, 12-5 and 1-6. However, didn't get flagged for behavioral data. Check
           '661', # Two files, see behavioral file
           '517', # Two files, 17-6 and 11-6. Not flagged in behavioral data. Check
          ]

def extract_video(filename: str, moment: str, date: str):
    if moment == 'A':
        file = data_A_dir + "\\" + filename
        processed_dir = data_A_dir + '\\video_features'
    elif moment == 'B':
        file = data_B_dir + "\\" + filename
        processed_dir = data_B_dir + '\\video_features'

    if not os.path.isfile(processed_dir + "\\" + filename[0:-5] + ".csv"): # Check if the file not yet exists to be savvy about computing power
        t = time.time()
        # Run OpenSmile
    #     subprocess.run(['C:\\Users\\mitch\\OneDrive - UGent\\UGent\\Topics\\Facial\\OpenFace_2.2.0_win_x64_new\\FeatureExtraction.exe',
    #                         '-f', file, '-out_dir', processed_dir, "-2Dfp",   "-3Dfp", "-pdmparams", "-pose", "-aus", "-gaze", "-hogalign"])
        subprocess.run(['D:\\UGent\\Topics\\Facial\\OpenFace_2.2.0_win_x64_new\\FeatureExtraction.exe',
                            '-f', file, '-out_dir', processed_dir, "-2Dfp",   "-3Dfp", "-pdmparams", "-pose", "-aus", "-gaze", "-hogalign"])
        if time.time() - t < 120: # check if it took less than 120 (this is an estimate of a good cut-off) seconds to analyze
                print("Very fast processing video features for: " + filename + ". Check whether everything has gone well!")
    
    

index = 0
for pptnum in tqdm(AVidID): # Loop through all file ID's
    if pptnum not in exclude: # Some participants messed up, so skip those. See earlier cell for descriptions of each
        index += 1
        indices = [i for i, s in enumerate(AVid) if pptnum in s] # Check per pptnum how many files are found
        if len(indices) > 1: # if more than one, something is going wrong, figure out
            print("Too many .webm files for ppt " + pptnum + ".")
            break
        elif len(indices) == 0: # if none, probably only saved before video, so check wCAM
            print("No .webm files found for ppt " + pptnum + ".")
        elif len(indices) == 1: # if one, that is perfect, read it
#             data = pd.read_csv(ACsv[indices[0]])
            filename = AVid[indices[0]]
            print("Currently processing video features for: " + filename)
            date = re.search(pptnum + '_(.*)webm', filename).group(1)
            extract_video(filename, 'A', date)

#         break
#     if index > 10:
#         break

t = time.time()
time.time() - t < 120

time.time() - t
